requirements.txt
websocket-client==1.6.1
rich==13.7.0
python-dotenv==1.0.0
numpy==1.24.3
scikit-learn==1.3.2
joblib==1.3.2
pandas==2.0.3
matplotlib==3.7.2
